requirements.txt
discord.py==1.3.3
async-timeout==2.0.1
twilio==6.38.1
cryptography==2.8
mysql-connector-python==8.0.19